optimize space filler, too
