+++ src/features/shapes/lsystem/LSystem.js
@@ -2,13 +2,13 @@ import Shape from "../Shape"
 import {
   lsystem,
   lsystemPath,
+  lsystemOptimize,
   onSubtypeChange,
   onMinIterations,
   onMaxIterations,
 } from "@/common/lindenmayer"
 import { subtypes } from "./subtypes"
-import { resizeVertices, cloneVertices } from "@/common/geometry"
-import { buildGraph, edgeKey } from "@/common/Graph"
+import { resizeVertices } from "@/common/geometry"
 
 const options = {
   subtype: {
@@ -64,65 +64,13 @@ export default class LSystem extends Shape {
       config.angle = Math.PI / 2
     }
 
-    const curve = lsystemPath(lsystem(config), config)
+    const path = lsystemOptimize(lsystemPath(lsystem(config), config), config)
     const scale = 18.0 // to normalize starting size
-    const path =
-      config.shortestPath >= iterations ? this.shortestPath(curve) : curve
 
     return resizeVertices(path, scale, scale)
   }
 
   getOptions() {
     return options
-  }
-
-  shortestPath(nodes) {
-    const graph = buildGraph(nodes)
-    const path = []
-    const visited = {}
-
-    for (let i = 0; i < nodes.length - 1; i++) {
-      const node1 = nodes[i]
-      const node2 = nodes[i + 1]
-      let node1Key = node1.toString()
-      let edge12Key = edgeKey(node1, node2)
-
-      if (visited[edge12Key]) {
-        const unvisitedNode = this.nearestUnvisitedNode(
-          i + 1,
-          nodes,
-          visited,
-          graph,
-        )
-
-        if (unvisitedNode != null) {
-          const shortestSubPath = graph.dijkstraShortestPath(
-            node1Key,
-            unvisitedNode.toString(),
-          )
-
-          path.push(...cloneVertices(shortestSubPath.slice(1)))
-          i = nodes.indexOf(unvisitedNode) - 1
-        }
-      } else {
-        path.push(node2)
-        visited[edge12Key] = true
-      }
-    }
-
-    return path
-  }
-
-  nearestUnvisitedNode(nodeIndex, nodes, visited, graph) {
-    for (let i = nodeIndex; i < nodes.length - 1; i++) {
-      const node1 = nodes[i]
-      const node2 = nodes[i + 1]
-
-      if (!visited[edgeKey(node1, node2)]) {
-        return node2
-      }
-    }
-
-    return null // all nodes visited
   }
 }
